+++ cmd/tplcfg.go
@@ -1,59 +1,61 @@
 package cmd
 
+import "github.com/gojekfarm/ziggurat-go/cmd/templates"
+
 var templateConfig = []ZigTemplate{
 	{
 		TemplateName:    "yaml-config",
-		TemplatePath:    "cmd/templates/config.yaml.tpl",
+		TemplateText:    templates.YamlConfig,
 		TemplateOutPath: "$APP_NAME/config/config.yaml",
 	},
 	{
 		TemplateName:    "go-mod",
-		TemplatePath:    "cmd/templates/go.mod.tpl",
+		TemplateText:    templates.GoMod,
 		TemplateOutPath: "$APP_NAME/go.mod",
 	},
 	{
 		TemplateName:    "main",
-		TemplatePath:    "cmd/templates/main.go.tpl",
+		TemplateText:    templates.Main,
 		TemplateOutPath: "$APP_NAME/main.go",
 	},
 	{
 		TemplateName:    "docker-compose-metrics.yaml",
-		TemplatePath:    "cmd/templates/docker-compose-metrics.yaml.tpl",
+		TemplateText:    templates.DockerComposeMetrics,
 		TemplateOutPath: "$APP_NAME/sandbox/docker-compose-metrics.yaml",
 	},
 	{
 		TemplateName:    "prom-conf",
-		TemplatePath:    "cmd/templates/prom_conf.yaml.tpl",
+		TemplateText:    templates.PromConf,
 		TemplateOutPath: "$APP_NAME/sandbox/prom_conf.yaml",
 	},
 	{
 		TemplateName:    "docker-compose-kafka",
-		TemplatePath:    "cmd/templates/docker-compose-kafka.yaml.tpl",
+		TemplateText:    templates.DockerComposeKafka,
 		TemplateOutPath: "$APP_NAME/sandbox/docker-compose-kafka.yaml",
 	},
 	{
 		TemplateName:    "makefile",
-		TemplatePath:    "cmd/templates/Makefile.tpl",
+		TemplateText:    templates.Makefile,
 		TemplateOutPath: "$APP_NAME/Makefile",
 	},
 	{
 		TemplateName:    "grafana-config",
-		TemplatePath:    "cmd/templates/grafana.ini.tpl",
+		TemplateText:    templates.GrafanaINI,
 		TemplateOutPath: "$APP_NAME/sandbox/grafana.ini",
 	},
 	{
 		TemplateName:    "telegraf-config",
-		TemplatePath:    "cmd/templates/telegraf.conf.tpl",
+		TemplateText:    templates.TelegrafConf,
 		TemplateOutPath: "$APP_NAME/sandbox/telegraf.conf",
 	},
 	{
 		TemplateName:    "datasource-grafana",
-		TemplatePath:    "cmd/templates/datasource.yaml.tpl",
+		TemplateText:    templates.GrafanaDS,
 		TemplateOutPath: "$APP_NAME/sandbox/datasource.yaml",
 	},
 	{
 		TemplateName:    "kafka-produce",
-		TemplatePath:    "cmd/templates/produce_messages.sh.tpl",
+		TemplateText:    templates.ProduceMessages,
 		TemplateOutPath: "$APP_NAME/sandbox/kafka_produce.sh",
 		IsExec:          true,
 	},