+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.6.1] 20202-11-06
+##Changes
+- fixes template not found bug
+- uses raw strings instead of tempalte files
+
 ## [v0.6.0] 20202-11-06
 ## Added
 - adds a sandbox to test out metric publishing,kafka production and retry flow